Changed linear regression in 'decoding_utils' to allow for complete set fitting when k_folds is 1. Added 'plot_single_cell_trace_by_task_variables' to interactively explore single neuron responses to combinations of current or prior choice and category.

+++ visualize_task_aligned_traces.py
@@ -116,7 +116,7 @@ if __name__ == '__main__':
         aligned_signal = np.zeros([interval_frames, len(state_start_frame)])
         for n in range(len(state_start_frame)):
             if np.isnan(state_start_frame[n]) == 0:
-               aligned_signal[:, n] = signal[current_neuron, state_start_frame[n] - int(interval_frames /2) : state_start_frame[n]  + int(interval_frames /2) + 1]
+               aligned_signal[:, n] = signal[current_neuron, int(state_start_frame[n]) - int(interval_frames /2) : int(state_start_frame[n])  + int(interval_frames /2) + 1]
             else: 
                aligned_signal[:,n] = np.nan
